Changed substage to input_reason
It turns out what I had been calling substage is actually
much better described as the reason why we are waiting for input.
the field is only non-zero when waiting_for_input, and the number
indicates what situation is being waited for.
So, I updated the simulator, the brewState type, and the decoder.

I also updated the logger to record the state when making brewstate
updates.